requirements.txt
fastapi
uvicorn
dlib
opencv-python-headless
numpy
imutils
Pillow
mtcnn
matplotlib
